Added option to never check sequence counts/lengths (to run faster).
Added command-line option -U/--never_check_readcounts_lengths - "Don't
check file readcounts/lengths at any stage (to save time) - note that
this means no by-stage readcount summary at the end!"
This should be useful when I just want to try some variant on the
processing, don't care that much about the details, and WANT THE DATA
FASTER (this takes a long time on HiSeq files!!)

I tested this on some real data, it works - the .fa/.fq output file is
the same with and without -U and with the old version of the program,
the _info.txt outfile is the same without -U as with the old version
(except headers/etc), and with -U it's just missing all the "# * output
file (*) data::" sections after each program (which contain
readcounts/sequence lengths) and the "### Final read count info" section
at the end.

Also minor bug-fix (info outfile formatting only): changed the "::" to
":" in the "# * output file (*) data::" lines.