New option to process mixed 5'/3' flanking regions; other stuff.
Split the -A/--full_cutadapt_options option into three parts:
-5/--adapter_5prime, -3/--adapter_3prime, and
-A/--other_cutadapt_options.  This makes it possible to process 5', 3',
and both-end files - apparently we'll be doing a lot of both-end files
in the future, so this is important.
At the cutadapt and later stages, all single outfiles are replaced with
a filename dictionary with 5' and 3' as keys (not all files must be
present).
Cutadapt is ran twice, once with each adapter (if both 5' and 3' are
given); each run generates cutadapt files and no-cassette files, and
afterward a real no-cassette file is generated from only the sequences
that show up in both no-cassette files (i.e. don't contain either
adapter).
The final output is modified to include the end info (5'/3') on all
"good read count" lines, and to include multiple such lines if we're
doing both ends.

Other stuff:
 - removed -U/--never_check_readcounts_lengths option - it was pretty
   useless, since it meant no summary at the end, and I need those for
   downstream stuff.
 - made sure all sequences were uppercase when written to fasta files,
   because some fastx_toolkit tools apparently choke on lowercase.
 - made all the little extra "### *" info bits get printed to the
   *_info.txt and not just to stdout
 - added comments/etc

Added new run-tests to test this 5'/3' split (test_data/pre__end*), and
renamed the old run-tests to make things clearer.
ALL RUN-TESTS PASS.