Bugfix: it works with no cutadapt now!
Previously it wouldn't generate a good-read outfile when cutadapt wasn't
used, because the no-cutadapt option changed the old cutadapt_tmpfile
instead of the new cutadapt_tmpfiles.  Fixed.

Also made it so when cutadapt isn't run, cutadapt_tmpfiles and
no_cassette_tmpfiles and outfiles aren't named with a 5prime/3prime (and
the dictionaries don't have 5prime/3prime as keys), because that
distinction only makes sense with cutadapt.

All run-tests work (although there is no run-test for this case);
running it on real data does now yield an output file when cutadapt
isn't used.