Added showing multi-aligned as unaligned SAM line with XM:i:M tag.
Implemented printing multiple-aligned cases as an unaligned SAM line
with an extra XM:i:M tag, where M is the number of detected alignments.
(Which is going to be 2, since for this output to happen, you have to
use the show-multiple 0 option, which results in running bowtie with the
minimum number of alignments to show, i.e. 2, since that still allows
differentiation between unique and multi.)

Added run-test case for that (new files); passes.

Changed unaligned_as_fasta default to False - seems safer, to avoid
mixing SAM and fasta.

Added some docstrings/comments.