Changed stdout to use length ranges.
Instead of reporting each alignment length (10-45), which is a lot of
numbers, report the less common lengths as ranges and the more common as
singles. (If you want singles, you can always do a uniq -c on the
cassette-info outfile)

Also changed some other stdout formatting to be nicer.

Adjusted unit-tests to take the new summary output; run-tests didn't
need adjustment.

Changed one run-test to not have the --quiet option, to leave the
formatting visible.